requirements.txt
pytest==8.3.4
playwright==1.51.0
pytest-playwright==0.5.0
pytest-xdist==3.5.0
allure-pytest==2.13.5
pytest-cov==5.0.0
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
